add uniffi bindings and set up compilation
