+++ crates/shielder_bindings/.gitignore
@@ -1,2 +1,3 @@
 pkg/
 artifacts/
+ios-bindings/
