+++ crates/shielder_bindings/uniffi-bindgen.rs
@@ -1,0 +1,4 @@
+// binary for generating uniffi language bindings
+fn main() {
+    uniffi::uniffi_bindgen_main()
+}
